# Faster Bi-Objective Non-Dominated Sorting with O(N log N) Complexity