@blankjul 2025-10-05
Amazing thanks for your contribution!

I was under the impression that this already exists somewhere in the codebase, but doesn't seem like it!