@MLopez-Ibanez 2025-10-06
Also, why would you want the file `test_comparison.py` in the root of the source directory and not under `tests/`? Is this a file that should be available to users of the package?