@blankjul 2025-11-24
all fair points. benchmark against the current implementation as well:
```
Bi-Objective Non-dominated Sorting Performance Benchmark
======================================================================
Size       Default (ms)    Native BiObj (ms)  Speedup    Correct
----------------------------------------------------------------------
50         0.107           0.036              2.98      x Yes
100        0.103           0.100              1.03      x Yes
500        1.038           1.065              0.98      x Yes
1000       3.205           3.094              1.04      x Yes
2000       8.615           8.642              1.00      x Yes
----------------------------------------------------------------------
All results correct: Yes
```

Decided not to make it default. 

and thanks for the callout. decided to make a benchmark folder within tests.