@MLopez-Ibanez 2025-11-28
> all fair points. benchmark against the current implementation as well:

I recently added `moocore` benchmarks for nondominated sorting: https://multi-objective.github.io/moocore/python/#nondominated-sorting-pareto-ranking

As you can see the current implementation of `pymoo`  is the slowest of the ones tested.